Clean Up Code & Add Battery Levels
- Removed unused functions to reduce size and complexity
- Added reading of wixel battery on P0_5
- Changed the formatting of the data output to match the expected values
in the DexDrip Android app (App was reading battery from the second
field [1] rather than third and raw data from [0].  Wixel was sending
data as RAW FILTERED BATT, with the filtered data never being used by
the android app.  Since three fields are expected by the Android app,
this puts the expected fields in their respective positions and allows
for the wixel battery voltage to be sent in the third field to be used
once I develop the integration in the app.
- Moved variables to XDATA if possible, reducing PSEG to 155